+++ src/Pdf/Service/FilenameCalculator.php
@@ -13,26 +13,13 @@ use FreshAdvance\Invoice\DataType\InvoiceDataInterface;
 
 class FilenameCalculator implements FilenameCalculatorInterface
 {
+    public function __construct(
+        private readonly FormatCalculatorInterface $formatCalculator,
+    ) {
+    }
+
     public function calculateByFormat(string $format, InvoiceDataInterface $invoiceData): string
     {
-        $order = $invoiceData->getOrder();
-
-        /** @var string $format */
-        $format = preg_replace_callback(
-            '/<order:(\w+)>/',
-            function ($matches) use ($order): string {
-                /** @var int|string|null $value */
-                $value = $order->getFieldData($matches[1]);
-                return (string)$value;
-            },
-            $format
-        );
-
-        $invoiceConfiguration = $invoiceData->getInvoiceConfiguration();
-        /** @var int|string|null $billNr */
-        $billNr = $order->getFieldData('oxbillnr');
-        $invoiceNumber = $invoiceConfiguration->getFormattedNumber((string)$billNr);
-
-        return str_replace('<invoiceNumber>', $invoiceNumber, $format);
+        return $this->formatCalculator->calculateByFormat($format, $invoiceData);
     }
 }