+++ tests/Unit/DataType/InvoiceConfigurationTest.php
@@ -33,19 +33,6 @@ class InvoiceConfigurationTest extends TestCase
         $this->assertSame('someNumber', $sut->getNumber());
     }
 
-    public function testGetFormattedNumber(): void
-    {
-        $sut = new InvoiceConfiguration(
-            orderId: 'someOrderId',
-            signer: 'someSigner',
-            date: 'someDate',
-            number: 'for%1$smat'
-        );
-
-        $invoiceNumber = uniqid();
-        $this->assertSame('for' . $invoiceNumber . 'mat', $sut->getFormattedNumber($invoiceNumber));
-    }
-
     #[DataProvider('formattedDateDataProvider')]
     public function testGetFormattedDate(string $format, string $expectation): void
     {